+++ src/app/modules/ngx-cron/ngx-cron.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NgxCronComponent } from './ngx-cron.component';
+import { CronService } from './ngx-cron.service';
 import { NgxUIModule } from '@swimlane/ngx-ui';
 
 @NgModule({
@@ -8,6 +9,7 @@ import { NgxUIModule } from '@swimlane/ngx-ui';
     CommonModule,
     NgxUIModule
   ],
+  providers: [CronService],
   declarations: [NgxCronComponent],
   exports: [NgxCronComponent]
 })