+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Feature: Component can be disabled
+* Added feature to disable quartz
 
 --------------------